* commit '7df6300d2b6c2a7268cbc2bba327693f89ee1bba':
  BE-17: set proper spec for test used income records
  BE-17: fulfill teacher income detail data
  BE-17: render income detail list daily

+++ server/app/models.py
@@ -1271,7 +1271,7 @@ class TimeSlot(BaseModel):
         return False
 
     def duration_hours(self):
-        return (self.end - self.start).seconds/3600
+        return int((self.end - self.start).seconds/3600)
 
     @property
     def trans_to_time(self):
